Corrects junit vulnerability with junit >= 4.7 and < 4.13.1
On Unix like systems, the system's temporary directory is shared between
all users on that system. Because of this, when files and directories
are written into this directory they are, by default, readable by other
users on that same system.

This vulnerability does not allow other users to overwrite the contents
of these directories or files. This is purely an information disclosure
vulnerability.